User Manager now with 100% test coverage

+++ src/public/classes/Data/UserManager.php
@@ -63,7 +63,8 @@ use DataTable\InMemoryDataTable;
  * 
  * @author Rafael Nájera <rafael.najera@uni-koeln.de>
  */
-class UserManager {
+class UserManager 
+{
     
     private $userTable;
     private $relationsTable;
@@ -83,7 +84,8 @@ class UserManager {
      * @param DataTable $ut
      * @param DataTable $rt
      */
-    public function __construct($ut = NULL, $rt = NULL, $pt = NULL) {
+    public function __construct($ut = NULL, $rt = NULL, $pt = NULL)
+    {
         $this->userTable = ($ut===NULL) ? new InMemoryDataTable() : $ut;
         $this->relationsTable = ($rt===NULL) ? new InMemoryDataTable() : $rt;
         $this->peopleTable = ($pt===NULL) ? new InMemoryDataTable() : $pt;
@@ -100,7 +102,8 @@ class UserManager {
      * @param int $userId
      * @return boolean
      */
-    public function userExistsById($userId){
+    public function userExistsById($userId)
+    {
         return $this->getUsernameFromUserId($userId) !== false;
     }
     
@@ -110,7 +113,8 @@ class UserManager {
      * @param string $userName
      * @return boolean
      */
-    public function userExistsByUserName($userName){
+    public function userExistsByUserName($userName)
+    {
         return $this->getUserIdFromUserName($userName) !== false;
     }
     
@@ -121,13 +125,12 @@ class UserManager {
      * @param int $userId
      * @return string
      */
-    public function getUsernameFromUserId($userId){
-        if ($this->userTable->rowExistsById($userId)){
+    public function getUsernameFromUserId($userId)
+    {
+        if ($this->userTable->rowExistsById($userId)) {
             return $this->userTable->getRow($userId)['username'];
         }
-        else{
-            return false;
-        }
+        return false;
     }
     
     /**
@@ -136,7 +139,8 @@ class UserManager {
      * @param string $userName
      * @return int
      */
-    public function getUserIdFromUserName($userName){
+    public function getUserIdFromUserName($userName)
+    {
         return $this->userTable->getIdForKeyValue('username', $userName);
     }
     
@@ -145,21 +149,24 @@ class UserManager {
      * @param int $userid User ID
      * @param array $userinfo Array where the information will be stored
      */
-    public function getUserInfoByUserId($userid){
-        
+    public function getUserInfoByUserId($userid)
+    {
         $pi = $this->peopleTable->getRow($userid);
         if ($pi=== false) {
             return false;
         }
         $ui = $this->userTable->getRow($userid);
         
-        if ($pi['email']){
+        if (!isset($pi['email'])) {
+            $pi['email'] = '';
+        }
+        if ($pi['email']) {
             // from https://en.gravatar.com/site/implement/hash/
             $emailhash =  md5(strtolower(trim($pi['email']))) ;
         } 
         else {
             $pi['email'] = '';
-            $emailhash =  'nohash';
+            $emailhash =  '';
         }
         
         
@@ -176,7 +183,7 @@ class UserManager {
         if ($fullName === '') {
             return false;
         }
-        if ($this->userExistsById($userId)){
+        if ($this->userExistsById($userId)) {
             $newInfo = [];
             $newInfo['id'] = $userId;
             $newInfo['fullname'] = $fullName;
@@ -187,16 +194,18 @@ class UserManager {
         return false;
     }
     
-    public function getUserInfoByUsername($username){
+    public function getUserInfoByUsername($username)
+    {
         $userid = $this->getUserIdFromUserName($username);
         return $this->getUserInfoByUserId($userid);
     }
     
-    public function getUserInfoForAllUsers(){
+    public function getUserInfoForAllUsers()
+    {
         $allUsers = $this->userTable->getAllRows();
-        $allUserInfo = array();
+        $allUserInfo = [];
         foreach($allUsers as $user){
-            array_push($allUserInfo, $this->getUserInfoByUserId($user['id']));
+            $allUserInfo[] = $this->getUserInfoByUserId($user['id']);
         }
         return $allUserInfo;
     }
@@ -211,8 +220,9 @@ class UserManager {
      * @param string $userName
      * @return int
      */
-    public function createUserByUsername($userName){
-        if ($this->userExistsByUserName($userName)){
+    public function createUserByUsername($userName)
+    {
+        if ($this->userExistsByUserName($userName)) {
             return false;
         }
         $personId = $this->createPerson();
@@ -226,7 +236,7 @@ class UserManager {
      */
     private function createPerson()
     {
-        return $this->peopleTable->createRow(['fullname' => 'No name']);
+        return $this->peopleTable->createRow(['fullname' => '']);
     }
     //
     // Allowed action methods
@@ -234,14 +244,17 @@ class UserManager {
     
      /**
      * Returns true if a user is allowed to do $action
-     * @param string $action, the action, normally a verb, e.g.: 'edit-other-users'
+     * @param string $action, the action, normally a verb, 
+     *                e.g.: 'edit-other-users'
      * @return boolean
      */
-    public function isUserAllowedTo($userId, $action){
+    public function isUserAllowedTo($userId, $action)
+    {
         if ($this->isRoot($userId)){
             return true;
         }
-        return $this->relationsTable->findRow(['userId' => $userId, 'relation' => 'isAllowed', 'attribute' => $action]) !== false;
+        return $this->relationsTable->findRow(['userId' => $userId, 
+            'relation' => 'isAllowed', 'attribute' => $action]) !== false;
     }
     
     /**
@@ -250,38 +263,47 @@ class UserManager {
      * @param String $action
      * @return boolean True if successful, false otherwise
      */
-    public function allowUserTo($userId, $action){
+    public function allowUserTo($userId, $action)
+    {
         if ($this->isRoot($userId)){
             return true;
         }
         if (!$this->userExistsById($userId)){
             return false;
         }
-        $this->relationsTable->createRow(['userId' => $userId, 'relation' => 'isAllowed', 'attribute' => $action]);
-        return true;
+        return $this->relationsTable->createRow(['userId' => $userId, 
+            'relation' => 'isAllowed', 'attribute' => $action]) !== false;
     }
     
-    public function disallowUserTo($userId, $action){
-        if ($this->isRoot($userId)){
+    public function disallowUserTo($userId, $action)
+    {
+        if (!$this->userExistsById($userId)) {
             return false;
         }
-        $key = $this->relationsTable->findRow(['userId' => $userId, 'relation' => 'isAllowed', 'attribute' => $action]);
+        if ($this->isRoot($userId)) {
+            return false;
+        }
+        $key = $this->relationsTable->findRow(['userId' => $userId, 
+            'relation' => 'isAllowed', 'attribute' => $action]);
         if ($key === false){
             return true;
         }
         return $this->relationsTable->deleteRow($key);
     }
     
-    public function isUserA($userId, $role){
+    public function isUserA($userId, $role)
+    {
         
-        return $this->relationsTable->findRow(['userId' => $userId, 'relation' => 'hasRole', 'attribute' => $role]) !== false;
+        return $this->relationsTable->findRow(['userId' => $userId, 
+            'relation' => 'hasRole', 'attribute' => $role]) !== false;
     }
     
     //
     // Role methods
     //
     
-    public function setUserRole($userId, $role){
+    public function setUserRole($userId, $role)
+    {
         
         if (!$this->userExistsById($userId)){
             return false;
@@ -292,33 +314,46 @@ class UserManager {
             return true;
         }
         
-        $this->relationsTable->createRow(['userId' => $userId, 'relation' => 'hasRole', 'attribute' => $role]);
+        $this->relationsTable->createRow(['userId' => $userId, 
+            'relation' => 'hasRole', 'attribute' => $role]);
         
         return true;
         
     }
     
-    public function revokeUserRole($userId, $role) {
-        $key = $this->relationsTable->findRow(['userId' => $userId, 'relation' => 'hasRole', 'attribute' => $role]);
+    public function revokeUserRole($userId, $role)
+    {
+        if (!$this->userExistsById($userId)){
+            return false;
+        }
+        // root cannot be revoked any role except the root role
+        if ($this->isRoot($userId) && $role !== $this->rootRole) {
+            return false;
+        }
+        $key = $this->relationsTable->findRow(['userId' => $userId, 
+            'relation' => 'hasRole', 'attribute' => $role]);
         if ($key === false){
             return true;
         }
         return $this->relationsTable->deleteRow($key);
-        
     }
     
     //
     // root role methods
     //
     
-    public function isRoot($userId){
+    public function isRoot($userId)
+    {
         return $this->isUserA($userId, $this->rootRole);
     }
-    public function makeRoot($userId){
+    
+    public function makeRoot($userId)
+    {
         return $this->setUserRole($userId, $this->rootRole);
     }
     
-    public function revokeRootStatus($userId){
+    public function revokeRootStatus($userId)
+    {
         return $this->revokeUserRole($userId, $this->rootRole);
     }
     
@@ -331,53 +366,55 @@ class UserManager {
      * @param int $userId
      * @return string
      */
-    public function getUserToken($userId){
+    public function getUserToken($userId)
+    {
         if ($this->userExistsById($userId)){
             $userRow = $this->userTable->getRow($userId);
             if (isset($userRow['token'])){
                 return $userRow['token'];
             }
-            else {
-                return '';
-            }
+            return '';
         }
         return false;
     }
     
-    public function storeUserToken($userId, $token){
+    public function storeUserToken($userId, $token)
+    {
         if ($this->userExistsById($userId)){
-            return $this->userTable->updateRow(['id' => $userId, 'token' => $token]);
+            return $this->userTable->updateRow(['id' => $userId, 
+                'token' => $token]);
         }
-        else {
-            return false;
-        }
+        return false;
     }
     
     // 
     // Passwords
     //
     
-    public function verifyUserPassword($userName, $givenPassword){
+    public function verifyUserPassword($userName, $givenPassword)
+    {
         if (!$this->userExistsByUserName($userName)){
             return false;
         }
         $u = $this->userTable->getRow($this->getUserIdFromUserName($userName));
-        if (!$u['password']){
+        if (!isset($u['password']) || $u['password'] === '') {
             return false;
         }
         //error_log("UM: Checking against hash " . $u['password']);
         return password_verify($givenPassword, $u['password']);
     }
     
-    public function storeUserPassword($userName, $password){
-        if ($password == '') {
+    public function storeUserPassword($userName, $password)
+    {
+        if ($password === '') {
             return false;
         }
         $hash = password_hash($password, PASSWORD_BCRYPT);
 
         if ($this->userExistsByUserName($userName)){
             $userId = $this->getUserIdFromUserName($userName);
-            return $this->userTable->updateRow(['id' => $userId, 'password' => $hash]);
+            return $this->userTable->updateRow(['id' => $userId, 
+                'password' => $hash]);
         }
         return false;
     }
